psf-scale-factor: image and PSF are cropped using Warp instead of Crop
Until now, the stamps of the objects and the PSF to make the calibration in
flux of the PSF was done using Crop. However, this caused some difficulties
related with small offsets when considering floating point pixels. It is
possible to use Warp instead of Crop for making this task. The difficulty
is just that maybe the image and/or the PSF do not have WCS information,
and thus, Warp is not able to do the job. In this situation it could be
possible to inject a fake WCS info and then use Warp.

With this this task has been added, although it is not finished yet. Now
this script uses Warp for making the crop of both the image and the PSF. In
the cases where no WCS information is available, a fake/artificial WCS
information is generated within the script to allow Warp making the warping
and obtain the desired image and PSF stamps so they can be divided to
obtain the scaling flux factor.

In addition to this, I also started to implement a new feature. It consists
in the possibility of considering different shifts around the given
coordinate of the star. This is done by two loops so the coordinates are
located around the center one. Then the ratio and the standard deviation
between the star and the PSF are computed within the specified ring. After
doing this, there will be a grid of flux factors, one for each shift. The
one that gives the minimum standard deviation value is the one considered
the best.

Why I started implementing this? Because it is interesting in those
situations when the center of the stars are not too accurate, or maybe the
star has moved because its intrinsic proper motion. There could be many
other situations but in general, having this feature will allow a better
accurate center estimation when modeling the stars using the PSF.