psf-scale-factor: estimating a better center position
Until now, there was some work related to shifting the coordinates to
estimate the best combination that reduces the standard deviation within
the ring for estimating the flux scaling factor. This commit is a
continuation in this field.

With this commit, I have included some options to be able to generalize the
estimation of a better center position. Still there are some things that
need to be added and better explained. But at least, right now it is
possible to use the script with the new features from the command line.