psf-scale-factor: checking that x and y-shifts inputs
Until now, the options provided to these parameters were not checked. As a
consquence, if the user provided wrong parameters, the script would crash
without providing useful information.

With this commit, checks and information has been provided. Now it checks
in advance that the parameters set are correct, and if not, it provides
useful information. In addition to this, more comments and information has
been added to the script.