psf-subtract: positioning the PSF image by Warp instead of Crop
Until now, the PSF was allocated into the right posision of the sky by
using Crop. However, computing the necessary shifts and translations
necessary for Crop was not easy.

With this commit, the way the PSF is warped into the  correct sky position
has changed. Now it is done by Warp. In the cases where no WCS information
are available from the input image, fake WCS information is used.  If the
input imate does contain WCS info, then it is used. The key idea is to
situate the center of the PSF image in the center of the object (star) that
is being subtracted.