psf-stamp: creating the stamps by Warp instead of Crop
Until now, the cropping of the image around the desired object was done by
Crop. However, computing the necessary shifts and translations for Crop was
not easy and sometimes it caused problems.

With this commit, the way the stamps are created has changed. Now it is
done by Warp. In the cases when no WCS information is available from the
input image, fake WCS information is used.  If the input image does contain
WCS info, then it is used. The key idea is to situate the center of the
cropped/warped image at the center of the object (star) that is being
considered with sub-pixel precision (and no shifts or offsets).