Warp: optimized by ignoring output pixels with no overlap
Until now, when an output pixel did not overlap with the input image in
linear warps, the program performed unneeded calculations.

With this commit, the output pixels that are fully outside of the input
image are detected and no further processing is done on them. This
optimization is able to halve the execution time of many troublesome
projections.

Note: the 'start > end' situation occurs in special projections. It is
probably due to a misbehaviour, and will be addressed in the future. Bug