Warp: remove unwanted implicit translation
Until now, when '--translate' was invoked in warp, it performed a "silent"
(1,1) translation. In addition to this, the size of the image is
incremented more than necessary.

With this commit, the operated translation is the intended one. This was
done by setting the minimum 'outfpixval' to (1,1), as in the
non-translational case, instead of (0,0). I also changed the correction to
'dsize' correction: in case of a decimal translation, a +1 already occur in
the main computation, so it is not needed. That +1 in the correction also
led to wrong computations in the case of a translation along only one axis.

This bug was reported by Raul Infante-Sainz.

This fixes bug #63026.