Crop: returns successfully when printing a single number
Until now, when the '--oneelemstdout' option was passed to crop, it would
print the value successfully, but return with a non-zero integer. This
happened because of the recent change in Commit 0ba08c1c27 (where we
checked for the existence of an output file to return successfully or not).
This issue was found while investigating the 'astscript-psf-stamp' script,
and I noticed that when it is not given the '--mode' option, it crashes.

With this commit, a new variable has been defined to keep track of such
cases (where the output is not a FITS file) and return successfully. The
problem with 'astscript-psf-stamp' was fixed by checking the calling of
'--mode' and by correcting the double quotations (in case the user
mistakenly gives a multi-word value).