Makefile plugins: text-prev functions when target not in list
Until now, when the 'ast-text-prev*' functions were given a target that
didn't exist in the input list, they would return the last batch of the
full list! This is wrong: when not in the input, the Make culture expects
an empty string.

With this commit, a check has been added to only continue with printing the
batch when the given target was actually found in the input list.