Makefile extensions: accounting for more RAM than needed
Until now, the 'ast-text-prev-batch-by-ram' Makefile extension would return
a "Floating point exception (core dumped)" error when its 'NEEDED_RAM_GB'
argument was larger than the available RAM on the machine.

With this commit, the problem has been fixed by adding a check for the
number of elements in each batch: in case it is zero (as it happened here),
it will become one (effectively no parallelization if used in a
prerequisite list). Also, a sanity check has been added at the start of the
worker function to catch such cases by other functions.