Makefile extensions: using fprintf(stderr,...) instead of error()
Until now, the Makefile extension functions used 'error()'. But this is a
GNU feature and within Gnuastro it is built internally by the library. As a
Makefile extension, it is 'make' which links to these functions and it may
have its own different implementation; producing segmentation fault (when
trying to find this function!). This was found when the number given to
'ast-text-prev-batch' was zero (causing a segmentation fault because of not
being able to call 'error').

With this commit, all the instances of 'error()' have been replaced with
'fprintf(); exit(1)'. Also, a check as been added so if the number given to
'ast-text-prev-batch' has a value of 0, Make aborts with an error message
to inform the user.