Statistics: new option to calculate concentration of distribution
Until now, there was no objective (independent of distribution's values)
way to measure the "concentration" of the given distribution around the
median. This is necessary in many contexts in addition to '--quantofmean'
(to make sure that the median and mean are the same and thus that the
median can be interpreted as the "center").

With this commit, a new function has been added for this purpose. The main
workhorse is in a newly added function in the library, which the Statistics
program uses.