Library (statistics.h): corrected typo in new concentration function
Until now, in the concentration function, we would only scale the input
values back to their original values when the operation was not done in
place! But this is wrong: it is only needed when we are doing the operation
in-place (the non-inplace scenario immediately frees the array). Also, the
command to correct it was just a copy to scale the input to the range of 0
to 1! So it was wrong also!

With this commit, both have been fixed.