astscript-fits-view: only using screen geometry when value was read
Until now, when the screen geometry could not be read, the value to DS9's
'-geometry' option would become an empty string, causing an error when
'astscript-fits-view' was called.

With this commit, to avoid a crash we now check the resulting string and if
it is empty, we ignore DS9's '-geometry' option.

This bug was reported by Ooldooz Kabood.

This fixes bug #65662.