Library: removed un-necessary GSL header includes
Until now, there were multiple instances of including GSL header files that
were not actually used within that file. Mostly just for historical
reasons: we originally had a function to do use that header, but that
function has moved to another file within Gnuastro.

With this commit, the un-necessary GSL header files that I could check in a
short time are removed (there may be more, but that needs more thorough
checks).