Library (arithmetic.h and dimension.h): maskfilled operators max area
Until now, when we were using the newly added maskfilled operators and one
of the input datasets was fully bad (for example a much higher noise level
than the rest), it because of filling holes, the outer edges of the bad
dataset would remain, while most of the inner region would be masked. This
would make the thin outer edges have different statistics than the main
area of the final stack!

With this commit, we now check the area covered by the mask and if it
covers more than 95% of the image's size, then the whole image is masked.