Arithmetic: correcting collapse-number to not use median
Until now, when the Arithmetic program was given the 'collapse-number'
operator, it would mistakenly call the internal macro that is used for
'collapse-median'! This originated in Commit 3aa904eca8c (from September
2022 where the sigma-clipping collapse operators were added). Furthermore,
while checking the sort-based collapse operators ('collapse-median' and all
the clip-based collapses), I noticed they are also giving segmentation
faults on an input which had a full row of NaN values.

With this commit, the first problem is fixed by fixing the typo (and giving
the correct macro is for 'collapse-number'). The second problem was caused
by the new feature of the blank-checking operator (that would free the
space when all elements were NaN). To avoid the problem, we now added a
check immediately after the sort-based operators and if the array is freed,
we'll allocate a new one for use in later iterations.

The original problem (with 'collapse-number') was reported by Sepideh
Eskandarlou.

This fixes bug #65743.