Library (pdf.h): using execl instead of system to avoid shell
Until now, when running Ghostscript to generate the PDF output of
ConvertType, we would use the 'system()' function of C. However, because
'system()' is generic, it uses the default '/bin/sh' shell to execute the
given command. This can cause problems when '/bin/sh' is not executable
(for example it has linking problems in the running environment).

With this commit, we are using the lower-level 'execl()' function of C to
avoid the shell overall, and directly pass our command to the system.

This was fixed as part of bug #65681, but it is not yet the full fix
because there are more 'system()' calls within Gnuastro that need to be
fixed.

This bug was reported by Sepideh Eskandarlou.