fits-view: adding support for DS9 region(s)
Until now, DS9 regions could either be loaded manually inside the DS9
window, or to give the DS9 commands to load regions with the '--ds9extra'
option of this script. But both were inconvenient.

With this commit, a new option ('--ds9region') is added to the script,
allowing to load any number of regions with a simple interface.