fits-view: fixed typo that didn't let some short options get parsed
Until now, some of the options would not work in their short format because
the removal of the short format string assumed the wrong letter!

With this commit, these typos are corrected.

Note from Mohammad while merging into 'master': a description of the newly
added '--ds9region' in the previous commit (8eb178da8) has also been added
to the 'NEWS' file in this commit.