radial-profile: added WCS to polar plot and bug fix in MakeProfile
Until now the polar plot did not have any WCS to better guide the users of
the output on what every pixel corresponds to. Also, while doing some
tests, I noticed a bug in MakeProfiles (it does not write negative values
when '--replace' is used).

With this commit, the first draft of the polar plot code from the previous
commit is edited and polished prior to merging with the 'master' branch and
the bug in MakeProfiles has been fixed.

This fixes bug #65822.