Convolve: correctly accounts for single element kernels
Until now, when a single-pixel kernel was given to Convolve, it would
crash by complaining about non-equal dimensions of inputs, even though the
single-pixel kernel was actually 2 dimensional!

With this commit, the root of the problem was fixed in the library function
'gal_dimension_remove_extra'. Also, the error message in Convolve has been
edited to be more clear.

This fixes bug #65833.