radial-profile: FITS polar plots correctly generated when output is text
Until now, when a plain-text output file was requested, the internally
created FITS file contents would be directly copied into it. But plain-text
editors do not recognize the FITS format, as a result, the file was not
readable.

With this commit, the problem has been fixed by checking if the output
format is FITS or not and properly setting the internal commands to copy
the final output from the internal one, while also copying the polar plot's
FITS file.