Statistics: --ontile operations done as expected
Until now, when any statistical operation was requested from Statistics
(for example in the command below), it would crash with a warning about not
being able to print one-row options with '--printparams'. This warning has
nothing to do with the '--ontile' option and was confusing!

With this commit, the cause of the problem was found to be the writing of
option values in the output FITS header keywords (which call the same
low-level function as '--printparams'). The problem was fixed by actually
printing a "1" for the called options in such cases (because only the
called options would be printed).

Also, as part of this commit, I noticed that the clipping operators were
not implemented with the '--ontile' option! So I added them.

This fixes bug #65853.