color-faint-gray: test in make check ignored when libjpeg not found
Until now, when libjpeg was not available, the 'mack check' test for
astscript-color-faint-gray would FAIL because its output is a JPEG file,
but it did not have a check for libjpeg.

With this commit, the necessary check has been added so this test is
correctly SKIPped instead of FAILing.

This bug was reported by Takashi Ichikawa.

This fixes bug #65847.