Statistics: directory of output respected with --ontile
Until now, when the Statistics program was called with '--oneline', and the
output was not in the running directory, the output's directory would be
ignored and the file is created in the running directory!

With this commit, the problem was solved by temporarily changing an
internal variable (the internal 'keepinputdir' which does not apply to this
scenario; as in the '--sky' opiton of Statistics).

This fixes bug #65862.