Code and Book: Corrected typos in comments and book text
Until now, there were some typos in the book and also in the comments
within the source code of the Table, NoiseChisel, Query programs, and
the 'Arithmetic.c' library. Also, the name of the 'ui_free_report'
function in Fits and Arithmetic programs differed from others.

With this commit, all these typos have been corrected. Also, for more
readability, the name of the 'ui_free_report' function has been unified in
all programs.