Library (arithmetic): new zero point changing operator
Until now, the only function that we had to convert zero points was to/from
nanomaggy (zero point of 22.5). But this value is not suitable in some
scenarios and other custom constant zero points may be necessary.

With this commit, a new 'zeropoint-change' operator has been added to the
arithmetic library (and thus the Arithmetic and Table programs) for this
job. In order to do it, a new function in 'units.h' has been added to do
the low-level operation.