Arithmetic: physical constants returned without crash
Until now, when Arithmetic was given a physical constant (like 'c' for the
speed of light; as in 'astarithmetic c'), it would crash by complaining
that the constant's code is not recognized.

With this commit, the problem was found to be in the absense of any
'break;' within the 'case' that searched the constants.

This fixes bug #65935.