Library (txt.h): col. num. correctly written in +100 col. tables
Until now, when a table to be written in plain-text had more than 100
columns, the first 9 column numbers would be incorrectly written in the
metadata. This happened because the buffer array that kept the column
number was not being initialized from the start.

With this commit, the problem was fixed by initializing the buffer array
that was used to write the column numbers.

This bug was found with the help of Sepideh Eskandarlou.

This fixes bug #65939.