Book: edited section on Gnuastro 1.0
Until now, this section had not been edited from the first time I wrote it
almost 9 years ago! At that time, Maneage wasn't yet created and the
description was not too clear (especially since Maneage is now a different
software package).

With this commit, the text in that section has been edited to be more clear
as to the goal of Gnuastro 1.0. The vision has not changed from before,
just the way to say it has become more precise/clear.