.gitignore: now ignoring TAGS files
Until now, we didn't explicitly ask Git to ignore TAGS files (which are
created by the 'etags' program that is part of Emacs). The TAGS help to
easily jump to function definitions within Emacs. But since they are
updated all the time, they should not be within the version controlled
source.

With this commit, the 'TAGS' file is automatically ignored by Git within
Gnuastro.