Book: sigclip-* operators account for the numbers image
Until now, some instances of the 'sigclip-*' operators of the Arithmetic
program didn't account for the fact that this operator now returns two
datasets (the previous output and the numbers image).

With this commit, after a search of this operator such instances were found
and the commands have been corrected.

This was reported by Hok Kan (Ronald) Tsang.