Book: explained how to derive SKY_STD in MakeCatalog's --sum-error
Until now, the description of the '--sum-error' column of MakeCatalog did
not mention how the users can create the sky standard deviation image.

With this commit, the description now has a link to NoiseChisel to clarify
this point. Two other minor issues in the radial profile script have been
implemented in this commit:
  1. The '--instd' and '--stdhdu' options were listed under the "Output"
     group of options. They have been brought to the "Input" group.
  2. The error message for the number of values given to the '--azimuth'
     option did not contain a new-line character in the end. As a result,
     the prompt of the user would be printed after it in interactive mode
     (which is inconvenient).

This was suggested by Rahna Payyasseri Thanduparackal.