MakeCatalog: new --mean-error option and re-write of std vs. std error
Until now, the section that described the difference between standard
deviation and standard error was confusing because it called the "standard
error" as "error" (which is the other name of standard deviation when
dealing with a normal distribution). Also, MakeCatalog didn't have any
column to measure the standard error, and there was no way to tell
MakeCatalog to avoid pixel values when measuring the error columns (when
the standard deviation image already contains the signal sources of error).

With this commit, all these issues are addressed: the title of the section
of the book was edited to be more clear and the text was also fully edited
and re-written where necessary. Also, the '--mean-error' option was added
to let people measure the standard error and the '--novalinerror' option
was defined so the pixel values are ignored when measuring errors.

The issue of the confusing text of that section in the book was raised by
Rahna Payyasseri Thanduparackal.