Book: corrected example of sigclip-* operators
Until now, the last example command of the 'sigclip-*' operators (under the
"Stacking operators" section of the book) contained the old
'madclip-fill-median' operator which was changed to the more
generic/powerful 'madclip-maskfilled' operator. This would cause a crash
when a reader tried the command with Gnuastro 0.23.

With this commit, the command has been corrected to use the operators that
are available at the moment.

This problem as reported by Alejandro Lumbreras-Calle.