Book: 3D cubes spectra includes section with logarithmic axises
Until now, this tutorial only described linear spectral coordinates
(because they are the way MUSE cubes are designed). But other surveys like
MaNGA provide logarithmic units in the third dimension.

With this commit, to simplify the usage of such cubes, a new short section
was added to this tutorial with a description and commands now how they can
identify such cubes and how they can plot the measured spectra.