astscript-color-faint-gray: manpage is installed
Until now, the Make rule to build the man-page for this installed script
was present, but it was never specified as a target to be installed! Ole
Streicher reported this problem, thanks to Debian's Lintian warning.

With this commit, it has been put on the list of targets to be installed
and distributed.