Library (fits.h): better grammar in keyword comments with versions
Until now, the comments of the FITS keywords that report the version of
Gnuastro's dependencies and Gnuastro itself was in the format of "Version
of AAAAA used." (where 'AAAAA' is the name of the software). But this way
of using 'used' makes it feel like a second-hand GSL for example!!!

With this commit, the the 'used' has been moved to after the software name
to properly convey the meaning.

This problem was reported by Joseph Putko.