Book: known issue added for "Numeric user ID too large" in make dist
Until now, in some operating systems and under some special circumnstances,
we were not able to create the Gnuastro distribution tarball with "make
dist".  This problem is related to Tar and how Automake calls it. For more
information see https://savannah.gnu.org/bugs/index.php?65578

With this commit, a manual work-around has been added to the "known issues"
section of the book so the users can generate the Gnuastro tarball until it
is fixed. This temporary fix is thanks to the discussions with Mohammad
Akhlaghi and Giacomo Lorenzetti.