MakeCatalog: setting --spatialresolution to 0 by default
Until this commit, the measurement of the surface brightness (SB) error
were not correct. After some checks I realized that this is because of the
use of '--spatialresolution' default value of 2. In Task #16586
(https://savannah.gnu.org/task/?16586) it can be seen that a better value
for this option is 0.

With this commit, I have set this option to 0 by default and a summary
description of the cause was added in the book.