astscript-color-faint-gray: The output of astmkprof was not set correctly
Until now, if we wanted to use a kernel and convolve the background, we
make a gussian kernel with astmkprof. But the output of the command was not
correct and it could not save the kernel.

With this commit, the correct variable was used for the output name. The
solution was also verified by Raul Infante-Sainz.

This fixes bug #66262.