Library (wcs.h): fix name issue when converting
Until now, when converting image coordinates to world coordinates (and
vice-versa) with `gal_wcs_img_to_world` (`gal_wcs_world_to_img`), the name
of the output columns (when 'inplace==0') were not the ones corresponding
to the actual data. This was caused by list initialized in reverse order.

With this commit, the list is created in the correct order.

This fixes bug #66308