Match: --kdtree is mandatory
Until now, if '--kdtree' was not set the program astmatch exited hinting a
bug, saying "a bug! [...] The code 0 isn't recognized for
'kdtreemode'". However, 0 is alias for "MATCH_KDTREE_INVALID", so the
option was de facto "mandatory".

With this commit, the 'kdtree' option's mandatory flag is activated. As a
result, the program will smoothly exit with "to continue, the following
options need a value" if a value is not given for it.

This fixes bug #66309