ConvertType: new --cmapblankcolor and --cmappgfplots
Until now, blank values would be displayed as the 0-value for the
particular colormap that was used. This could be confusing in scenarios
where 0 is meaningful in the image. Also, to use ConvertType's PDFs with
PGFPlots for high-quality LaTeX plots, we were forced to use colormas that
we had in common with it (only 'gray' and 'viridis').

With this commit, both issues have been solved through the newly added
options: '--cmapblankcolor' allows setting any color (which should ideally
be outside the color-space of the colormap) and '--cmappgfplots' produces a
PGFPlots colormap definition that can be easily '\input' into LaTeX to use
any Gnuastro colorbar.

The '--cmapblankcolor' option was suggested by Sepideh Eskandarlou.