Match: now has multiple arrangements of output rows
Until now, Match only done one type of output arrangement: what is known in
SQL as the INNER JOIN. But other arrangements are also sometimes necessary.

With this commit, the other well-known arragements of an output of Match
have been implemented: full, outer and outer-within-aperture. The exapmles
in the newly added section of the book fully describe them.