Book: fix incorrect description of AND operator
Until now, the documentation of the AND operator was incomplete. It
literally stated that if both operands of the AND operator are zero, then
the result is zero. This is technically correct, but it leaves the reader
wondering about the generic case when one operand is zero and the other is
non-zero.

With this commit, the wording is changed to clarify that if either are
zero, the returned value is 0. This is the conventional definition of the
AND operator and is also the way that the AND operator functions in
Gnuastro.