Library (options.h): --cite now includes book (using Zenodo DOI)
Until now, in the citation outputs of the programs we had didn't guide the
users on how to cite the book! Also, there was no clear explanation on why
the NoiseChisel paper is suggested as the main citation!

With this commit, the proper Zenodo DOI for this version of Gnuastro is
being used for a BibTeX entry that is recommended to the users. Also, some
explanation has been added for the main Gnuastro paper.

This was done after a good discussion with Rahna Payyasseri Thanduparackal.