Book: typos in docs and tutorial of GAL_DIMENSION_NEIGHBOR_OP
Until now, we had forgot to include the 'error.h' header in the library
demo section that showed the usage of this function, and also there was a
typo in the description of it ('ndim' was written as 'ndim'!)

With this commit, both are fixed. Also, to help better find the library
demo section from the description of the macro, another link to it has been
added after describing the operation.