astscript-psf-*: using the correct centering position for the stars/PSF
Until now, the correct center position of the PSF has been not properly
taken into account. Actually, there are few commits with some changes
because the results were not good at all. There was always 1-pixel offsets.

With this commit, after several tests and checks, the center position of
the PSF is properly taken into account both in the estimation of the flux
scale factor and also in the subtraction of the PSF itself.

This fixes bug #66043.