astscript-radial-profile: center after crop found more robustly
Until now, the radial profile script done a relatively complex, not
commented and hard to follow process to convert the original center in each
coordinate into a center for the cropped image.

With this commit, a more robust way to do the conversion has been
implemented which uses the fact that Gnuastro's crop will ensure that the
original center coordiante is within its central pixel. Furthermore a good
comment has been written to clarify the logic behind the process.