Crop: logfile name is based on output name
Until now, the log-file of Crop was always called 'astcrop.log' and it was
a plain-text file. This would cause problems when running in parallel and
was inefficient (see the NEWS file edit of this commit for more). Also, we
were mistakenly setting the table to keep track of how many images were
made based on the number of input images, not output images. As a result a
segmentation fault could be created.

With this commit, both issues have been addressed: the log file is now a
FITS file and its name is based on the output name. Also, the table to keep
track of the number of crops is now based on the number of input rows.

This bug was reported by Raul Infante-Sainz.

This fixes bug #66869.