Book: added example for Query's --overlapwith optimization
Until now, there was no example to guide the user on how to only keep
sources that we actually have data for in the '--overlapwith' option of
Query. Also, while reviewing some of the code, I noticed some parts of the
table program had lines longer than 75 characters and there was a wrongly
placed set of empty lines at the start of 'type.h'.

With this commit, a complete set of commands have been added to demonstrate
how to limit the output of Query's '--overlapwith' to those that are within
the given input image. Both the stylistic issues have also been addressed
with this commit.

This issue was raised by Raul Infante-Sainz.

This addresses the issue raised in bug #66867.